AT.2
Versão para entrega do AT

+++ AT/src/main/java/br/edu/infnet/spring_boot/AT/controller/DepartamentoController.java
@@ -1,6 +1,7 @@
 package br.edu.infnet.spring_boot.AT.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -22,26 +23,33 @@ public class DepartamentoController {
 
     @GetMapping("/departamento")
     public ResponseEntity<?> getDepartamentos(){
-        return departamentoService.getDepartamentos();
+        if(departamentoService.getDepartamentos().isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não existem departamentos!");
+        return ResponseEntity.ok(departamentoService.getDepartamentos());
     }
 
     @GetMapping("/departamento/{id}")
     public ResponseEntity<?> getDepartamentoById(@PathVariable Long id){
-        return departamentoService.getDepartamentoById(id);
+        if(departamentoService.getDepartamentoById(id).isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Departamento não existente!");
+        return ResponseEntity.ok(departamentoService.getDepartamentoById(id));
     }    
 
     @PostMapping("/departamento")
     public ResponseEntity<?> createCursoByAlunoId(@RequestBody Departamento departamento) {
-        return departamentoService.createDepartamento(departamento);
+        if(departamentoService.createDepartamento(departamento) == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível cadastrar Departamento!");
+        return ResponseEntity.status(HttpStatus.CREATED).body("Departamento cadastrado!");        
     }
     
     @PutMapping("departamento/{id}")
     public ResponseEntity<?> updateCurso(@PathVariable Long id, @RequestBody Departamento departamento){
-        return departamentoService.updateDepartamento(id, departamento);
+        Departamento update = departamentoService.updateDepartamento(id, departamento);
+        if(update == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Departamento não existente!");
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Departamento atualizado!");
     }
 
     @DeleteMapping("/departamento/{id}")
     public ResponseEntity<?> deleteCursoById(@PathVariable Long id){
-        return departamentoService.deleteDepartamentoById(id);
+        if(departamentoService.getDepartamentoById(id).isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Departamento não existente!");
+        departamentoService.deleteDepartamentoById(id);
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Departamento excluído!");
     }    
 }
