+++ AT/src/main/java/br/edu/infnet/spring_boot/AT/controller/FuncionarioController.java
@@ -6,6 +6,7 @@ import br.edu.infnet.spring_boot.AT.model.Funcionario;
 import br.edu.infnet.spring_boot.AT.service.FuncionarioService;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,31 +22,39 @@ public class FuncionarioController {
     
     @GetMapping("/funcionario")
     public ResponseEntity<?> getFuncionarios(){
-        return funcionarioService.getFuncionarios();
+        if(funcionarioService.getFuncionarios().isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não existem funcionários!");
+        return ResponseEntity.ok(funcionarioService.getFuncionarios());
     }
 
     @GetMapping("/funcionario/{id}")
     public ResponseEntity<?> getFuncionarioById(@PathVariable Long id){
-        return funcionarioService.getFuncionarioById(id);
+        if(funcionarioService.getFuncionarioById(id).isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Funcionário não cadastrado!");
+        return ResponseEntity.ok(funcionarioService.getFuncionarioById(id));
     }
 
     @GetMapping("/funcionario/departamento/{idDepartamento}")
     public ResponseEntity<?> getFuncionariosByDepartamentoId(@PathVariable Long idDepartamento) {
-        return funcionarioService.getFuncionariosByDepartamentoId(idDepartamento);
+        if(funcionarioService.getFuncionarioById(idDepartamento).isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não existem funcionários neste departamento!");
+        return ResponseEntity.ok(funcionarioService.getFuncionariosByDepartamentoId(idDepartamento));
     }
     
     @PostMapping("/funcionario")
     public ResponseEntity<?> createcreateFuncionarioAluno(@RequestBody Funcionario funcionario) {
-        return funcionarioService.createFuncionario(funcionario);
+        if(funcionarioService.createFuncionario(funcionario) == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível cadastrar Usuário!");
+        return ResponseEntity.status(HttpStatus.CREATED).body("Funcionário cadastrado!");
     }
 
     @PutMapping("funcionario/{id}")
     public ResponseEntity<?> updateFuncionario(@PathVariable Long id, @RequestBody Funcionario funcionario){
-        return funcionarioService.updateFuncionario(id, funcionario);
+        Funcionario update = funcionarioService.updateFuncionario(id, funcionario);
+        if(update == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Funcionário não cadastrado!");
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Funcionário atualizado!");        
     }
 
     @DeleteMapping("/funcionario/{id}")
     public ResponseEntity<?> deleteFuncionarioById(@PathVariable Long id){
-        return funcionarioService.deleteFuncionarioById(id);
+        if(funcionarioService.getFuncionarioById(id).isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Funcionário não cadastrado!");
+        funcionarioService.deleteFuncionarioById(id);
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Funcionário excluído!");
     }
 }
