+++ AT/src/main/java/br/edu/infnet/spring_boot/AT/controller/UsuarioController.java
@@ -1,7 +1,9 @@
 package br.edu.infnet.spring_boot.AT.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,26 +22,35 @@ public class UsuarioController {
     
     @GetMapping("/usuario")
     public ResponseEntity<?> getUsuarios(){
-        return usuarioService.getUsuarios();
+        if(usuarioService.getUsuarios().isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não existem usuários!");
+        return ResponseEntity.ok(usuarioService.getUsuarios());
     }
 
     @GetMapping("/usuario/{id}")
     public ResponseEntity<?> getUsuarioById(@PathVariable String id){
-        return usuarioService.getUsuarioById(id);
+        if(usuarioService.getUsuarioById(id).isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado!");
+        return ResponseEntity.ok(usuarioService.getUsuarioById(id));
     }    
 
     @PostMapping("/usuario")
     public ResponseEntity<?> createUsuario(@RequestBody Usuario usuario) {
-        return usuarioService.createUsuario(usuario);
+        usuario.setSenha(new BCryptPasswordEncoder().encode(usuario.getSenha()));
+        if(usuarioService.createUsuario(usuario) == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível cadastrar Usuário!");
+        return ResponseEntity.status(HttpStatus.CREATED).body("Usuário cadastrado!");
     }
 
     @PutMapping("usuario/{id}")
     public ResponseEntity<?> updateUsuario(@PathVariable String id, @RequestBody Usuario usuario){
-        return usuarioService.updateUsuario(id, usuario);
+        usuario.setSenha(new BCryptPasswordEncoder().encode(usuario.getSenha()));
+        Usuario update = usuarioService.updateUsuario(id, usuario);
+        if(update == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado!");
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Usuário atualizado!");
     }
 
     @DeleteMapping("/usuario/{id}")
     public ResponseEntity<?> deleteUsuario(@PathVariable String id){
-        return usuarioService.deleteUsuario(id);
+        if(usuarioService.getUsuarioById(id).isEmpty()) return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado!");
+        usuarioService.deleteUsuario(id);
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body("Usuário excluído!");
     }
 }
